Align with the beryllium platform matrix
- Remove UWP builds
- .NET 7 MAUI -> .NET 8 MAUI
- Cleanup the service class a bit (auto register a few things that can be auto registered)
