+++ src/Couchbase.Lite.Shared/Couchbase.Lite.Shared.projitems
@@ -200,12 +200,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)Support\SerialQueue.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Support\Status.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Support\ThreadSafety.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Support\UWP\DefaultDirectoryResolver.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Support\UWP\MainThreadTaskScheduler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Support\UWP\Reachability.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Support\UWP\UwpConsoleLogger.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Support\UWP\UWPProxy.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Support\UWP\UwpRuntimePlatform.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Support\WinUI\DefaultDirectoryResolver.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Support\WinUI\Reachability.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Support\WinUI\WinUIConsoleLogger.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Support\WinUI\WinUIProxy.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Support\WinUI\WinUIRuntimePlatform.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Sync\AuthOptionsDictionary.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Sync\HTTPLogic.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Sync\HttpMessageParser.cs" />
