+++ src/Couchbase.Lite.Shared/Support/Android/XamarinAndroidProxy.cs
@@ -27,6 +27,7 @@ using Java.Net;
 
 namespace Couchbase.Lite.Support
 {
+    [CouchbaseDependency]
     internal sealed class XamarinAndroidProxy : IProxy
     {
         #region IProxy
