+++ src/Couchbase.Lite.Support.Apple/iOS/Couchbase.Lite.Support.iOS.csproj
@@ -3,8 +3,8 @@
   <PropertyGroup>
     <Configurations>Debug;Release;Packaging</Configurations>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <TargetFrameworks>net7.0-ios;net7.0-maccatalyst;Xamarin.iOS10</TargetFrameworks>
-    <SupportedOSPlatformVersion Condition="$(TargetFramework.StartsWith('net7')) and $([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">12.0</SupportedOSPlatformVersion>
+    <TargetFrameworks>net8.0-ios;net8.0-maccatalyst;Xamarin.iOS10</TargetFrameworks>
+    <SupportedOSPlatformVersion Condition="$(TargetFramework.StartsWith('net8')) and $([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'ios'">12.0</SupportedOSPlatformVersion>
     <SupportedOSPlatformVersion Condition="$([MSBuild]::GetTargetPlatformIdentifier('$(TargetFramework)')) == 'maccatalyst'">13.1</SupportedOSPlatformVersion>
     <OutputType>Library</OutputType>
     <RootNamespace>Couchbase.Lite.Support.iOS</RootNamespace>
@@ -47,7 +47,7 @@
     <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
     <DocumentationFile>bin\Packaging\Couchbase.Lite.Support.iOS.xml</DocumentationFile>
   </PropertyGroup>
-  <PropertyGroup Condition="$(TargetFramework.StartsWith('net7.0-ios')) or $(TargetFramework.StartsWith('net7.0-maccatalyst'))">
+  <PropertyGroup Condition="$(TargetFramework.StartsWith('net8.0-ios')) or $(TargetFramework.StartsWith('net8.0-maccatalyst'))">
     <DefineConstants>$(DefineConstants);NET_APPLE</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
