--- src/Couchbase.Lite.Support.UWP/Couchbase.Lite.Support.UWP.csproj
@@ -1,119 +1,0 @@
-﻿<Project Sdk="MSBuild.Sdk.Extras">
-  <PropertyGroup>
-    <Configurations>Debug;Release;Packaging;PackagingDebug</Configurations>
-    <Platforms>x64;arm64</Platforms>
-    <TargetFramework>uap10.0.19041</TargetFramework>
-    <OutputType>Library</OutputType>
-    <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Couchbase.Lite.Support.UWP</RootNamespace>
-    <AssemblyName>Couchbase.Lite.Support.UWP</AssemblyName>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
-    <TargetPlatformVersion Condition=" '$(TargetPlatformVersion)' == '' ">10.0.19041.0</TargetPlatformVersion>
-    <TargetPlatformMinVersion>10.0.19041.0</TargetPlatformMinVersion>
-    <MinimumVisualStudioVersion>16</MinimumVisualStudioVersion>
-    <FileAlignment>512</FileAlignment>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
-    <CblEnterprisePackage Condition="'$(CblEnterprisePackage)' == ''">false</CblEnterprisePackage>
-    <PackageId Condition="'$(CblEnterprisePackage)' == 'true'">Couchbase.Lite.Enterprise.Support.UWP</PackageId>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Debug'">
-    <PlatformTarget>$(Platform)</PlatformTarget>
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE;NETFX_CORE;WINDOWS_UWP</DefineConstants>
-    <NoWarn>;2008</NoWarn>
-    <DebugType>portable</DebugType>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <ErrorReport>prompt</ErrorReport>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Release'">
-    <PlatformTarget>$(Platform)</PlatformTarget>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE;NETFX_CORE;WINDOWS_UWP</DefineConstants>
-    <Optimize>true</Optimize>
-    <NoWarn>;2008</NoWarn>
-    <DebugType>portable</DebugType>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <ErrorReport>prompt</ErrorReport>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'Packaging'">
-    <PlatformTarget>$(Platform)</PlatformTarget>
-    <OutputPath>bin\Packaging\</OutputPath>
-    <DefineConstants>TRACE;NETFX_CORE;WINDOWS_UWP</DefineConstants>
-    <Optimize>true</Optimize>
-    <NoWarn>;2008</NoWarn>
-    <NoStdLib>true</NoStdLib>
-    <DebugType>portable</DebugType>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <ErrorReport>prompt</ErrorReport>
-    <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
-    <DocumentationFile>bin\Packaging\Couchbase.Lite.Support.UWP.xml</DocumentationFile>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)' == 'PackagingDebug'">
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\PackagingDebug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE;NETFX_CORE;WINDOWS_UWP</DefineConstants>
-    <NoWarn>;2008</NoWarn>
-    <NoStdLib>true</NoStdLib>
-    <DebugType>portable</DebugType>
-    <PlatformTarget>$(Platform)</PlatformTarget>
-    <UseVSHostingProcess>false</UseVSHostingProcess>
-    <ErrorReport>prompt</ErrorReport>
-    <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
-  </PropertyGroup>
-  <PropertyGroup>
-    <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="Activate.cs" />
-  </ItemGroup>
-  <ItemGroup Condition="$(Configuration.Contains('Packaging'))">
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\x86_64-store\bin\LiteCore.dll">
-      <Link>x64\LiteCore.dll</Link>
-      <Pack>true</Pack>
-      <PackagePath>runtimes/win-x64/nativeassets/uap10.0/LiteCore.dll</PackagePath>
-    </None>
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\x86_64-store\bin\LiteCore.pdb">
-      <Link>x64\LiteCore.pdb</Link>
-      <Pack>true</Pack>
-      <PackagePath>runtimes/win-x64/nativeassets/uap10.0/LiteCore.pdb</PackagePath>
-    </None>
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\arm64-store\bin\LiteCore.dll">
-      <Link>arm64\LiteCore.dll</Link>
-      <Pack>true</Pack>
-      <PackagePath>runtimes/win-arm64/nativeassets/uap10.0/LiteCore.dll</PackagePath>
-    </None>
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\arm64-store\bin\LiteCore.pdb">
-      <Link>arm64\LiteCore.pdb</Link>
-      <Pack>true</Pack>
-      <PackagePath>runtimes/win-arm64/nativeassets/uap10.0/LiteCore.pdb</PackagePath>
-    </None>
-  </ItemGroup>
-  <ItemGroup Condition="!$(Configuration.Contains('Packaging'))">
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\x86_64-store\bin\LiteCore.dll">
-      <Link>LiteCore.dll</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\x86_64-store\bin\LiteCore.pdb">
-      <Link>LiteCore.pdb</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\arm64-store\bin\LiteCore.dll">
-      <Link>LiteCore.dll</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="$(MSBuildThisFileDirectory)..\..\vendor\prebuilt_core\windows\arm64-store\bin\LiteCore.pdb">
-      <Link>LiteCore.pdb</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-  <Import Project="..\AssemblyInfo.props" />
-</Project>
