+++ src/Couchbase.Lite.Support.WinUI/Couchbase.Lite.Support.WinUI.csproj
@@ -2,7 +2,7 @@
   <PropertyGroup>
     <Configurations>Debug;Release;Packaging;PackagingDebug</Configurations>
     <Platforms>x64;arm64</Platforms>
-    <TargetFramework>net6.0-windows10.0.19041.0</TargetFramework>
+    <TargetFramework>net8.0-windows10.0.19041.0</TargetFramework>
     <RuntimeIdentifiers>win-x64;win-arm64</RuntimeIdentifiers>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
@@ -20,7 +20,7 @@
     <PlatformTarget>$(Platform)</PlatformTarget>
     <DebugSymbols>true</DebugSymbols>
     <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE;NETFX_CORE;NET6_0_WINDOWS10</DefineConstants>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
     <NoWarn>;2008</NoWarn>
     <DebugType>portable</DebugType>
     <UseVSHostingProcess>false</UseVSHostingProcess>
@@ -29,7 +29,7 @@
   <PropertyGroup Condition="'$(Configuration)' == 'Release'">
     <PlatformTarget>$(Platform)</PlatformTarget>
     <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>TRACE;NETFX_CORE;NET6_0_WINDOWS10</DefineConstants>
+    <DefineConstants>TRACE</DefineConstants>
     <Optimize>true</Optimize>
     <NoWarn>;2008</NoWarn>
     <DebugType>portable</DebugType>
@@ -39,7 +39,7 @@
   <PropertyGroup Condition="'$(Configuration)' == 'Packaging'">
     <PlatformTarget>$(Platform)</PlatformTarget>
     <OutputPath>bin\Packaging\</OutputPath>
-    <DefineConstants>TRACE;NETFX_CORE;NET6_0_WINDOWS10</DefineConstants>
+    <DefineConstants>TRACE</DefineConstants>
     <Optimize>true</Optimize>
     <NoWarn>;2008</NoWarn>
     <NoStdLib>true</NoStdLib>
@@ -53,7 +53,7 @@
     <DebugSymbols>true</DebugSymbols>
     <PlatformTarget>$(Platform)</PlatformTarget>
     <OutputPath>bin\PackagingDebug\</OutputPath>
-    <DefineConstants>DEBUG;TRACE;NETFX_CORE;NET6_0_WINDOWS10</DefineConstants>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
     <NoWarn>;2008</NoWarn>
     <NoStdLib>true</NoStdLib>
     <DebugType>portable</DebugType>
