--- src/Couchbase.Lite.Tests.UWP/Assets/C/tests/data/nested.json
@@ -1,4 +1,0 @@
-{"shapes": [{"color": "red", "size": 3}, {"color": "green", "size": 2}, {"color": "blue", "size": 10}]}
-{"shapes": []}
-{"shapes": [{"color": "red", "size": 8}, {"color": "yellow", "size": 5}]}
-{"shapes": [{"color": "cyan", "size": 3}, {"color": "white", "size": 1}]}
