--- src/Couchbase.Lite.Tests.UWP/Assets/C/tests/data/paths.json
@@ -1,3 +1,0 @@
-{"paths":[{"city":"San Francisco"},{"city":"Palo Alto"},{"city":"San Jose"}]}
-{"paths":[{"city":"Mountain View"},{"city":"Palo Alto"},{"city":"Belmont"}]}
-{"paths":[{"city":"San Francisco"},{"city":"Redwood City"},{"city":"San Mateo"}]}
