--- src/Couchbase.Lite.Tests.UWP/Assets/C/tests/data/sentences.json
@@ -1,36 +1,0 @@
-{"sentence": "It was a dark and stormy night; the rain fell in torrents, except at occasional intervals, when it was checked by a violent gust of wind which swept up the streets (for it is in London that our scene lies), rattling along the house-tops, and fiercely agitating the scanty flame of the lamps that struggled against the darkness."}
-{"sentence": "Through one of the obscurest quarters of London, and among haunts little loved by the gentlemen of the police, a man, evidently of the lowest orders, was wending his solitary way."}
-{"sentence": "He stopped twice or thrice at different shops and houses of a description correspondent with the appearance of the quartier in which they were situated, and tended inquiry for some article or another which did not seem easily to be met with."}
-{"sentence": "All the answers he received were couched in the negative; and as he turned from each door he muttered to himself, in no very elegant phraseology, his disappointment and discontent."}
-{"sentence": "At length, at one house, the landlord, a sturdy butcher, after rendering the same reply the inquirer had hitherto received, added, “But if this vill do as vell, Dummie, it is quite at your sarvice!”"}
-{"sentence": "Pausing reflectively for a moment, Dummie responded that he thought the thing proffered might do as well; and thrusting it into his ample pocket, he strode away with as rapid a motion as the wind and the rain would allow."}
-{"sentence": "He soon came to a nest of low and dingy buildings, at the entrance to which, in half-effaced characters, was written “Thames Court.”"}
-{"sentence": "Halting at the most conspicuous of these buildings, an inn or alehouse, through the half-closed windows of which blazed out in ruddy comfort the beams of the hospitable hearth, he knocked hastily at the door."}
-{"sentence": "He was admitted by a lady of a certain age, and endowed with a comely rotundity of face and person."}
-{"sentence": "“Hast got it, Dummie?” said she, quickly, as she closed the door on the guest."}
-{"sentence": "“Noa, noa! not exactly; but I thinks as ‘ow—”"}
-{"sentence": "“Pish, you fool!” cried the woman, interrupting him peevishly."}
-{"sentence": "“Vy, it is no use desaving me."}
-{"sentence": "You knows you has only stepped from my boosing-ken to another, and you has not been arter the book at all."}
-{"sentence": "So there’s the poor cretur a raving and a dying, and you—”"}
-{"sentence": "“Let I speak!” interrupted Dummie in his turn."}
-{"sentence": "“I tells you I vent first to Mother Bussblone’s, who, I knows, chops the whiners morning and evening to the young ladies, and I axes there for a Bible; and she says, says she, ‘I’ as only a “Companion to the Halter,” but you’ll get a Bible, I think, at Master Talkins’, the cobbler as preaches.’"}
-{"sentence": "So I goes to Master Talkins, and he says, says he, ‘I ‘as no call for the Bible,—‘cause vy? I ‘as a call vithout; but mayhap you’ll be a getting it at the butcher’s hover the vay,—‘cause vy? The butcher ‘ll be damned!’"}
-{"sentence": "So I goes hover the vay, and the butcher says, says he, ‘I ‘as not a Bible, but I ‘as a book of plays bound for all the vorld just like ‘un, and mayhap the poor cretur may n’t see the difference.’"}
-{"sentence": "So I takes the plays, Mrs. Margery, and here they be surely!"}
-{"sentence": "And how’s poor Judy?”"}
-{"sentence": "“Fearsome! she’ll not be over the night, I’m a thinking.”"}
-{"sentence": "“Vell, I’ll track up the dancers!”"}
-{"sentence": "So saying, Dummie ascended a doorless staircase, across the entrance of which a blanket, stretched angularly from the wall to the chimney, afforded a kind of screen; and presently he stood within a chamber which the dark and painful genius of Crabbe might have delighted to portray."}
-{"sentence": "The walls were whitewashed, and at sundry places strange figures and grotesque characters had been traced by some mirthful inmate, in such sable outline as the end of a smoked stick or the edge of a piece of charcoal is wont to produce."}
-{"sentence": "The wan and flickering light afforded by a farthing candle gave a sort of grimness and menace to these achievements of pictorial art, especially as they more than once received embellishments from portraits of Satan such as he is accustomed to be drawn."}
-{"sentence": "A low fire burned gloomily in the sooty grate, and on the hob hissed “the still small voice” of an iron kettle."}
-{"sentence": "On a round deal table were two vials, a cracked cup, a broken spoon of some dull metal, and upon two or three mutilated chairs were scattered various articles of female attire."}
-{"sentence": "On another table, placed below a high, narrow, shutterless casement (athwart which, instead of a curtain, a checked apron had been loosely hung, and now waved fitfully to and fro in the gusts of wind that made easy ingress through many a chink and cranny), were a looking-glass, sundry appliances of the toilet, a box of coarse rouge, a few ornaments of more show than value, and a watch, the regular and calm click of which produced that indescribably painful feeling which, we fear, many of our readers who have heard the sound in a sick-chamber can easily recall."}
-{"sentence": "A large tester-bed stood opposite to this table, and the looking-glass partially reflected curtains of a faded stripe, and ever and anon (as the position of the sufferer followed the restless emotion of a disordered mind) glimpses of the face of one on whom Death was rapidly hastening."}
-{"sentence": "Beside this bed now stood Dummie, a small, thin man dressed in a tattered plush jerkin, from which the rain-drops slowly dripped, and with a thin, yellow, cunning physiognomy grotesquely hideous in feature, but not positively villanous in expression."}
-{"sentence": "On the other side of the bed stood a little boy of about three years old, dressed as if belonging to the better classes, although the garb was somewhat tattered and discoloured."}
-{"sentence": "The poor child trembled violently, and evidently looked with a feeling of relief on the entrance of Dummie."}
-{"sentence": "And now there slowly, and with many a phthisical sigh, heaved towards the foot of the bed the heavy frame of the woman who had accosted Dummie below, and had followed him, haud passibus aequis, to the room of the sufferer; she stood with a bottle of medicine in her hand, shaking its contents up and down, and with a kindly yet timid compassion spread over a countenance crimsoned with habitual libations."}
-{"sentence": "This made the scene,—save that on a chair by the bedside lay a profusion of long, glossy, golden ringlets, which had been cut from the head of the sufferer when the fever had begun to mount upwards, but which, with a jealousy that portrayed the darling littleness of a vain heart, she had seized and insisted on retaining near her; and save that, by the fire, perfectly inattentive to the event about to take place within the chamber, and to which we of the biped race attach so awful an importance, lay a large gray cat, curled in a ball, and dozing with half-shut eyes, and ears that now and then denoted, by a gentle inflection, the jar of a louder or nearer sound than usual upon her lethargic senses."}
-{"sentence": "The dying woman did not at first attend to the entrance either of Dummie or the female at the foot of the bed, but she turned herself round towards the child, and grasping his arm fiercely, she drew him towards her, and gazed on his terrified features with a look in which exhaustion and an exceeding wanness of complexion were even horribly contrasted by the glare and energy of delirium."}
