--- src/Couchbase.Lite.Tests.UWP/Assets/localhost-wrong.cert
@@ -1,18 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIC+zCCAeOgAwIBAgIJAKnTwXlOTuqUMA0GCSqGSIb3DQEBBQUAMBQxEjAQBgNV
-BAMMCWxvY2FsaG9zdDAeFw0xNzA2MjMyMzI0MjlaFw0yNzA2MjEyMzI0MjlaMBQx
-EjAQBgNVBAMMCWxvY2FsaG9zdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoC
-ggEBAKLPmI5ZQXCkLBWakqf0BpYi1YvdLzM1IiJRSLx4lht5Ms1baDkaZ+CC8sC/
-6LedaSbRfc9ymqSRunWcKiuopr0gC5eiGOJuWwGtqNsbyqnLCBxaagbzUEUx7klY
-q0XKcaNJNE/gyGEAjqtNM4+l0wqmAiub/e3cStCI8nkDFJ3NmN/lT96seCxPMvFi
-VIlW+S0QnY9saMMl9LSq839keHalq1hJuh7VKuvv6HcfUsVsMI8obEgna8jhJfa2
-bwRL7fm8UAIZlUcIIt6cCzVUueyhsw4sa/lr3327KNMoWdyru0XTwfOEQA4+nkZv
-d4FikgXR8R/BczJ7Fctpye32Ve0CAwEAAaNQME4wHQYDVR0OBBYEFK3IlJjvRX+R
-F/DdwFNfogaBJjGkMB8GA1UdIwQYMBaAFK3IlJjvRX+RF/DdwFNfogaBJjGkMAwG
-A1UdEwQFMAMBAf8wDQYJKoZIhvcNAQEFBQADggEBAD4LnocV7ywVLEO7w/mIHe4d
-Y6BsjSsccZSQipUJgXNuxCSxsm0JUg2WvHSymaWFVsdwPbKXKHKtsEqf0CVxx7AA
-5/0xTT7xCmwTQUfX6/cRkOzjjDnKe+bianK1QklMhy0k7KILdLhh57YQE4IR6Uqy
-vv+rRf0vlPGUketBrMnBZdhDV07oz+WeYWNoPtdrJsGWBuXigcqoQV3q4ADzRTCc
-KyxHXljP3FBSKTI10VJ84z3r4C3f5VJklhmbhDZAOdPJiWahYJr77v1NMomndrlp
-0z6ouajAOpB1qC4MtuIULldsPABZ7TgMuig1mP2XMLc33f284tH/lkS+xVHbtw8=
------END CERTIFICATE-----
