--- src/Couchbase.Lite.Tests.UWP/UnitTestApp.xaml
@@ -1,8 +1,0 @@
-<Application
-    x:Class="Couchbase.Lite.Tests.UWP.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Couchbase.Lite.Tests.UWP"
-    RequestedTheme="Light">
-
-</Application>
