+++ src/Couchbase.Lite/Properties/AssemblyInfo.cs
@@ -1,10 +1,8 @@
 ﻿using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("Couchbase.Lite.Tests.UWP")]
 [assembly: InternalsVisibleTo("Couchbase.Lite.Tests.iOS")]
 [assembly: InternalsVisibleTo("Couchbase.Lite.Support.Android")]
 [assembly: InternalsVisibleTo("Couchbase.Lite.Tests.Android")]
 [assembly: InternalsVisibleTo("Couchbase.Lite.Tests.NetCore")]
 [assembly: InternalsVisibleTo("Couchbase.Lite.Tests.Maui")]
 [assembly: InternalsVisibleTo("Couchbase.Lite.Tests.NetCore.Validation")]
-[assembly: InternalsVisibleTo("Couchbase.Lite.Tests.UWP.Validation")]
